svg/struct/scripted/svg-getIntersectionList-004.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml"><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><script>const test = async_test(&apos;getIntersectionList() with nested &lt;svg&gt;&apos;); window.onload = test.step_func_done(() =&gt; { const testSVG = document.getElementById(&apos;askedForIntersection&apos;); const r = testSVG.createSVGRect(); r.x = 0; r.y = 0; r.width = 100; r.height = 100; assert_equals(testSVG.getIntersectionList(r, testSVG).length, 2); });</script><rect id="r1" width="100" height="100" x="0" y="0"/><svg id="askedForIntersection"><rect id="r2" width="100" height="100" x="0" y="0"/><svg id="nestedSVG"><rect id="r3" width="100" height="100" x="0" y="0"/></svg></svg></svg>
